Splitted CoursesController: CoursesUpdateController created (no modules or namespaces)
No changes to folder structure
courses_update_controller_spec added
