+++ app/controllers/courses_controller.rb
@@ -6,9 +6,9 @@ require "#{Rails.root}/lib/wiki_course_edits"
 require "#{Rails.root}/lib/list_course_manager"
 require "#{Rails.root}/lib/tag_manager"
 require "#{Rails.root}/lib/course_creation_manager"
-require "#{Rails.root}/app/workers/update_course_worker"
+# require "#{Rails.root}/app/workers/update_course_worker"
 require "#{Rails.root}/app/workers/notify_untrained_users_worker"
-require "#{Rails.root}/app/workers/announce_course_worker"
+# require "#{Rails.root}/app/workers/announce_course_worker"
 
 #= Controller for course functionality
 class CoursesController < ApplicationController
@@ -40,16 +40,16 @@ class CoursesController < ApplicationController
     @course = course_creation_manager.create
   end
 
-  def update
-    validate
-    handle_course_announcement(@course.instructors.first)
-    slug_from_params if should_set_slug?
-    @course.update course: course_params
-    set_timeline_enabled
-    ensure_passcode_set
-    UpdateCourseWorker.schedule_edits(course: @course, editing_user: current_user)
-    render json: { course: @course }
-  end
+  # def update
+  #   validate
+  #   handle_course_announcement(@course.instructors.first)
+  #   slug_from_params if should_set_slug?
+  #   @course.update course: course_params
+  #   set_timeline_enabled
+  #   ensure_passcode_set
+  #   UpdateCourseWorker.schedule_edits(course: @course, editing_user: current_user)
+  #   render json: { course: @course }
+  # end
 
   def destroy
     validate
@@ -158,33 +158,33 @@ class CoursesController < ApplicationController
     return unless user_signed_in? && current_user.instructor?(@course)
   end
 
-  def handle_course_announcement(instructor)
-    # Course announcements aren't particularly necessary, but we'll keep them on
-    # for Wiki Ed for now.
-    return unless Features.wiki_ed?
-    newly_submitted = !@course.submitted? && course_params[:submitted] == true
-    return unless newly_submitted
-    CourseSubmissionMailerWorker.schedule_email(@course, instructor)
-    AnnounceCourseWorker.schedule_announcement(course: @course,
-                                               editing_user: current_user,
-                                               instructor: instructor)
-  end
+  # def handle_course_announcement(instructor)
+  #   # Course announcements aren't particularly necessary, but we'll keep them on
+  #   # for Wiki Ed for now.
+  #   return unless Features.wiki_ed?
+  #   newly_submitted = !@course.submitted? && course_params[:submitted] == true
+  #   return unless newly_submitted
+  #   CourseSubmissionMailerWorker.schedule_email(@course, instructor)
+  #   AnnounceCourseWorker.schedule_announcement(course: @course,
+  #                                              editing_user: current_user,
+  #                                              instructor: instructor)
+  # end
 
   def should_set_slug?
     %i(title school).all? { |key| params[:course].key?(key) }
   end
 
-  def slug_from_params(course = params[:course])
-    slug = String.new("#{course[:school]}/#{course[:title]}")
-    slug << "_(#{course[:term]})" unless course[:term].blank?
+  # def slug_from_params(course = params[:course])
+  #   slug = String.new("#{course[:school]}/#{course[:title]}")
+  #   slug << "_(#{course[:term]})" unless course[:term].blank?
 
-    course[:slug] = slug.tr(' ', '_')
-  end
+  #   course[:slug] = slug.tr(' ', '_')
+  # end
 
-  def ensure_passcode_set
-    return unless course_params[:passcode].nil?
-    @course.update_attribute(:passcode, Course.generate_passcode)
-  end
+  # def ensure_passcode_set
+  #   return unless course_params[:passcode].nil?
+  #   @course.update_attribute(:passcode, Course.generate_passcode)
+  # end
 
   def initial_campaign_params
     params
@@ -198,16 +198,16 @@ class CoursesController < ApplicationController
       .permit(:language, :project)
   end
 
-  def set_timeline_enabled
-    case params.dig(:course, :timeline_enabled)
-    when true
-      @course.flags[:timeline_enabled] = true
-      @course.save
-    when false
-      @course.flags[:timeline_enabled] = false
-      @course.save
-    end
-  end
+  # def set_timeline_enabled
+  #   case params.dig(:course, :timeline_enabled)
+  #   when true
+  #     @course.flags[:timeline_enabled] = true
+  #     @course.save
+  #   when false
+  #     @course.flags[:timeline_enabled] = false
+  #     @course.save
+  #   end
+  # end
 
   def course_params
     params
