+++ config/environments/development.rb
@@ -11,7 +11,7 @@ Rails.application.configure do
   config.eager_load = false
 
   # Show full error reports and disable caching.
-  config.consider_all_requests_local       = false
+  config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
   config.action_mailer.default_url_options = { host: ENV["SMTP_HOST"] }
