+++ config/routes.rb
@@ -48,6 +48,11 @@ Rails.application.routes.draw do
       constraints: { course_id: /.*/ }
 
   # Courses
+  controller :courses_update do
+    put 'courses/:id' => 'courses_update#update', constraints: { id: /.*/ },
+        :as => :courses
+  end
+
   controller :courses do
     get 'courses/new' => 'courses#new',
         constraints: { id: /.*/ } # repeat of resources
@@ -86,7 +91,7 @@ Rails.application.routes.draw do
   get 'lookups/article(.:format)' => 'lookups#article'
 
   # Timeline
-  resources :courses, constraints: { id: /.*/ } do
+  resources :courses, except: :update, constraints: { id: /.*/ } do
     resources :weeks, only: [:index, :new, :create], constraints: { id: /.*/ }
     # get 'courses' => 'courses#index'
   end
